# apps/api/src/modules/ledger/categories.controller.ts
import {
  Controller,
  Post,
  Body,
  Patch,
  Param,
  Delete,
  UseGuards,
} from '@nestjs/common';
import { ApiTags, ApiOperation, ApiBearerAuth } from '@nestjs/swagger';
import { JwtAuthGuard } from '../../auth/guards/jwt-auth.guard';
import { CurrentUser } from '../../common/decorators/current-user.decorator';
import { CategoriesService } from './categories.service';
import { CreateCategoryDto } from './dto/create-category.dto';
import { UpdateCategoryDto } from './dto/update-category.dto';
import { BudgetShareGuard } from './budget-share.guard';

@ApiTags('categories')
@ApiBearerAuth()
@Controller('v1/categories')
@UseGuards(JwtAuthGuard, BudgetShareGuard)
export class CategoriesController {
  constructor(private readonly categoriesService: CategoriesService) {}

  @Post()
  @ApiOperation({ summary: 'Create a new category' })
  async createCategory(
    @CurrentUser() user: any,
    @Body() dto: CreateCategoryDto,
  ) {
    return this.categoriesService.createCategory(user.id, dto);
  }

  @Patch(':id')
  @ApiOperation({ summary: 'Update a category' })
  async updateCategory(
    @CurrentUser() user: any,
    @Param('id') id: string,
    @Body() dto: UpdateCategoryDto,
  ) {
    return this.categoriesService.updateCategory(user.id, id, dto);
  }

  @Delete(':id')
  @ApiOperation({ summary: 'Delete a category' })
  async removeCategory(@CurrentUser() user: any, @Param('id') id: string) {
    return this.categoriesService.removeCategory(user.id, id);
  }
}